# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ApiError:
    properties:
      code:
        type: string
      details:
        additionalProperties: true
        type: object
      message:
        type: string
    type: object
  dto.ApiResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/dto.ApiError'
      meta:
        $ref: '#/definitions/dto.ResponseMeta'
      success:
        type: boolean
    type: object
...
  dto.BulkPriorityRequest:
    properties:
      includeBreakdown:
        type: boolean
      taskIds:
        items:
          type: string
        maxItems: 100
        minItems: 1
        type: array
    required:
    - taskIds
    type: object
  dto.BulkPriorityResponse:
    properties:
      priorities:
        items:
          $ref: '#/definitions/dto.TaskPrioritySummary'
        type: array
      sortedOrder:
        items:
          type: string
        type: array
    type: object
...
  dto.GenerateNudgesRequest:
    properties:
      async:
        type: boolean
      projectId:
        type: string
      scope:
        enum:
        - project
        - organization
        type: string
      types:
        items:
          type: string
        type: array
    required:
    - scope
    type: object
...
  dto.IDResponse:
    properties:
      id:
        type: string
    type: object
...
  dto.NudgeAction:
    properties:
      action:
        type: string
      timestamp:
        type: string
      userId:
        type: string
    type: object
  dto.NudgeActionRequest:
    properties:
      actionType:
        enum:
        - accept_suggestion
        - dismiss
        - custom_action
        - ask_alternatives
        - snooze
        type: string
      parameters:
        additionalProperties: true
        type: object
    required:
    - actionType
    type: object
  dto.NudgeActionResponse:
    properties:
      actionTaken:
        type: string
      completedAt:
        type: string
      followUpNudges:
        items:
          type: string
        type: array
      nudgeId:
        type: string
      nudgeStatus:
        type: string
      result:
        $ref: '#/definitions/dto.NudgeActionResult'
    type: object
  dto.NudgeActionResult:
    properties:
      fromUserId:
        type: string
      taskId:
        type: string
      taskReassigned:
        type: boolean
      toUserId:
        type: string
    type: object
  dto.NudgeDetailResponse:
    properties:
      aiExplanation:
        type: string
      createdAt:
        type: string
      criticalityScore:
        type: integer
      description:
        type: string
      expiresAt:
        type: string
      history:
        items:
          $ref: '#/definitions/dto.NudgeAction'
        type: array
      id:
        type: string
      metrics:
        $ref: '#/definitions/dto.NudgeMetrics'
      relatedEntities:
        $ref: '#/definitions/dto.RelatedEntities'
      severity:
        type: string
      status:
        type: string
      suggestedAction:
        $ref: '#/definitions/dto.SuggestedAction'
      title:
        type: string
      type:
        type: string
    type: object
  dto.NudgeListResponse:
    properties:
      nudges:
        items:
          $ref: '#/definitions/dto.NudgeResponse'
        type: array
      summary:
        $ref: '#/definitions/dto.NudgeSummary'
    type: object
  dto.NudgeMetrics:
    properties:
      allocationPercentage:
        type: integer
      assignedTasks:
        type: integer
      totalHours:
        type: number
    type: object
  dto.NudgeResponse:
    properties:
      aiExplanation:
        type: string
      createdAt:
        type: string
      criticalityScore:
        type: integer
      description:
        type: string
      expiresAt:
        type: string
      id:
        type: string
      metrics:
        $ref: '#/definitions/dto.NudgeMetrics'
      relatedEntities:
        $ref: '#/definitions/dto.RelatedEntities'
      severity:
        type: string
      status:
        type: string
      suggestedAction:
        $ref: '#/definitions/dto.SuggestedAction'
      title:
        type: string
      type:
        type: string
    type: object
  dto.NudgeStatsResponse:
    properties:
      acted:
        type: integer
      actionRate:
        type: number
      avgTimeToAction:
        type: string
      byType:
        additionalProperties:
          $ref: '#/definitions/dto.NudgeTypeStats'
        type: object
      dismissed:
        type: integer
      expired:
        type: integer
      generated:
        type: integer
      period:
        type: string
    type: object
  dto.NudgeSummary:
    properties:
      bySeverity:
        additionalProperties:
          type: integer
        type: object
      byType:
        additionalProperties:
          type: integer
        type: object
      total:
        type: integer
      unread:
        type: integer
    type: object
  dto.NudgeTypeStats:
    properties:
      acted:
        type: integer
      generated:
        type: integer
    type: object
...
  dto.PriorityBreakdown:
    properties:
      businessValue:
        type: integer
      criticalPathWeight:
        type: integer
      deadlineUrgency:
        type: integer
      dependencyImpact:
        type: integer
      projectPriority:
        type: integer
    type: object
  dto.PriorityFactors:
    properties:
      blockedTasksCount:
        type: integer
      businessValue:
        type: integer
      daysUntilDue:
        type: integer
      deadlineUrgency:
        type: integer
      isOnCriticalPath:
        type: boolean
      projectPriority:
        type: integer
    type: object
...
  dto.ProjectTaskRankingResponse:
    properties:
      projectId:
        type: string
      rankings:
        items:
          $ref: '#/definitions/dto.TaskRankingItem'
        type: array
      total:
        type: integer
    type: object
...
  dto.RecalculatePriorityAsyncResponse:
    properties:
      estimatedDuration:
        type: string
      jobId:
        type: string
      status:
        type: string
    type: object
  dto.RecalculatePriorityRequest:
    properties:
      async:
        type: boolean
      projectId:
        type: string
      scope:
        enum:
        - project
        - organization
        - task
        type: string
    required:
    - scope
    type: object
  dto.RecalculatePrioritySyncResponse:
    properties:
      affectedTasks:
        items:
          type: string
        type: array
      duration:
        type: string
      recalculated:
        type: integer
    type: object
...
  dto.RelatedEntities:
    properties:
      personId:
        type: string
      projectId:
        type: string
      taskId:
        type: string
    type: object
...
  dto.ResponseMeta:
    properties:
      hasMore:
        type: boolean
      nextCursor:
        type: string
      page:
        type: integer
      perPage:
        type: integer
      requestId:
        type: string
      timestamp:
        type: string
      total:
        type: integer
    type: object
...
  dto.SuggestedAction:
    properties:
      description:
        type: string
      suggestedAssigneeId:
        type: string
      targetTaskId:
        type: string
      type:
        type: string
    type: object
...
  dto.TaskPriorityResponse:
    properties:
      breakdown:
        $ref: '#/definitions/dto.PriorityBreakdown'
      calculatedAt:
        type: string
      factors:
        $ref: '#/definitions/dto.PriorityFactors'
      rank:
        type: integer
      taskId:
        type: string
      universalPriorityScore:
        type: integer
    type: object
  dto.TaskPrioritySummary:
    properties:
      rank:
        type: integer
      taskId:
        type: string
      universalPriorityScore:
        type: integer
    type: object
...
  dto.TaskRankingItem:
    properties:
      assigneeId:
        type: string
      priorityScore:
        type: integer
      rank:
        type: integer
      status:
        type: string
      taskId:
        type: string
      title:
        type: string
    type: object
...
  dto.UpdateNudgeStatusRequest:
    properties:
      status:
        enum:
        - unread
        - read
        - dismissed
        - acted
        type: string
    required:
    - status
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Xephyr Project Management Intelligence API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Xephyr API
  version: "1.0"
paths:
...
  /nudges:
    get:
      consumes:
      - application/json
      description: Get a list of nudges with optional filters
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by severity
        in: query
        name: severity
        type: string
      - description: Filter by type
        in: query
        name: type
        type: string
      - description: Filter by project
        in: query
        name: projectId
        type: string
      - description: Filter by person
        in: query
        name: personId
        type: string
      - default: 20
        description: Limit results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.NudgeListResponse'
                meta:
                  $ref: '#/definitions/dto.ResponseMeta'
              type: object
      security:
      - BearerAuth: []
      summary: List nudges
      tags:
      - nudges
  /nudges/{nudgeId}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific nudge
      parameters:
      - description: Nudge ID
        in: path
        name: nudgeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.NudgeDetailResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get single nudge
      tags:
      - nudges
  /nudges/{nudgeId}/actions:
    post:
      consumes:
      - application/json
      description: Perform an action on a nudge (accept, dismiss, etc.)
      parameters:
      - description: Nudge ID
        in: path
        name: nudgeId
        required: true
        type: string
      - description: Action request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.NudgeActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.NudgeActionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Take action on nudge
      tags:
      - nudges
  /nudges/{nudgeId}/status:
    patch:
      consumes:
      - application/json
      description: Update the status of a nudge
      parameters:
      - description: Nudge ID
        in: path
        name: nudgeId
        required: true
        type: string
      - description: Status update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateNudgeStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.NudgeResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Update nudge status
      tags:
      - nudges
  /nudges/generate:
    post:
      consumes:
      - application/json
      description: Trigger manual nudge generation
      parameters:
      - description: Generation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GenerateNudgesRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.IDResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Generate nudges
      tags:
      - nudges
  /nudges/stats:
    get:
      consumes:
      - application/json
      description: Get statistics about nudges
      parameters:
      - default: 30d
        description: Time period
        in: query
        name: period
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.NudgeStatsResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get nudge statistics
      tags:
      - nudges
  /priorities/projects/{projectId}/ranking:
    get:
      consumes:
      - application/json
      description: Get prioritized task ranking for a specific project
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by assignee
        in: query
        name: assigneeId
        type: string
      - description: Minimum priority score
        in: query
        name: minScore
        type: integer
      - default: 50
        description: Max results (default 50)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Pagination offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ProjectTaskRankingResponse'
                meta:
                  $ref: '#/definitions/dto.ResponseMeta'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get task ranking for project
      tags:
      - priorities
  /priorities/recalculate:
    post:
      consumes:
      - application/json
      description: Trigger recalculation of priorities for a scope
      parameters:
      - description: Recalculation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RecalculatePriorityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RecalculatePrioritySyncResponse'
              type: object
        "202":
          description: Accepted
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RecalculatePriorityAsyncResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Trigger priority recalculation
      tags:
      - priorities
  /priorities/tasks/{taskId}:
    get:
      consumes:
      - application/json
      description: Get the priority score and breakdown for a specific task
      parameters:
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.TaskPriorityResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get single task priority
      tags:
      - priorities
  /priorities/tasks/bulk:
    post:
      consumes:
      - application/json
      description: Get priority scores for multiple tasks in a single request
      parameters:
      - description: Bulk priority request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.BulkPriorityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.BulkPriorityResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get priorities for multiple tasks
      tags:
      - priorities
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"